added editable headers & dynamic project pages
